Moving Contentful helpers into separate file.
