+++ resources/assets/components/ContentfulEntry/ContentfulEntry.js
@@ -2,17 +2,18 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import Loader from '../utilities/Loader';
+import { report, withoutNulls } from '../../helpers';
 import SoftEdgeBlock from '../actions/SoftEdgeBlock';
 import EmbedBlock from '../blocks/EmbedBlock/EmbedBlock';
 import ErrorBlock from '../blocks/ErrorBlock/ErrorBlock';
 import LandingPage from '../pages/LandingPage/LandingPage';
 import ImagesBlock from '../blocks/ImagesBlock/ImagesBlock';
+import { parseContentfulType } from '../../helpers/contentful';
 import ContentBlock from '../blocks/ContentBlock/ContentBlock';
 import GalleryBlock from '../blocks/GalleryBlock/GalleryBlock';
 import SectionBlock from '../blocks/SectionBlock/SectionBlock';
 import CampaignUpdate from '../blocks/CampaignUpdate/CampaignUpdate';
 import AffirmationContainer from '../Affirmation/AffirmationContainer';
-import { parseContentfulType, report, withoutNulls } from '../../helpers';
 import ActionStatsBlock from '../blocks/ActionStatsBlock/ActionStatsBlock';
 import CurrentClubBlock from '../blocks/CurrentClubBlock/CurrentClubBlock';
 import LinkActionContainer from '../actions/LinkAction/LinkActionContainer';
