+++ resources/assets/components/utilities/CuratedPageBanner.js
@@ -2,9 +2,10 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { css } from '@emotion/core';
 
-import StatCard, { STAT_PROPS } from './StatCard/StatCard';
+import { tailwind } from '../../helpers';
 import TextContent from './TextContent/TextContent';
-import { contentfulImageUrl, tailwind } from '../../helpers';
+import StatCard, { STAT_PROPS } from './StatCard/StatCard';
+import { contentfulImageUrl } from '../../helpers/contentful';
 
 const CuratedPageBanner = ({
   coverImage,
