+++ resources/assets/components/utilities/Gallery/templates/ContentBlockGalleryItem.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 
 import { Figure } from '../../Figure/Figure';
 import TextContent from '../../TextContent/TextContent';
-import { contentfulImageUrl } from '../../../../helpers';
+import { contentfulImageUrl } from '../../../../helpers/contentful';
 
 const ContentBlockGalleryItem = ({
   showcaseTitle,
