+++ resources/assets/components/utilities/Person/templates/AdvisoryBoardMemberTemplate/AdvisoryBoardMemberTemplate.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 
 import { Figure } from '../../../Figure/Figure';
 import TextContent from '../../../TextContent/TextContent';
-import { contentfulImageUrl } from '../../../../../helpers';
+import { contentfulImageUrl } from '../../../../../helpers/contentful';
 
 const AdvisoryBoardMemberTemplate = props => {
   const { showcaseTitle, showcaseImage, showcaseDescription } = props;
